Update bookmarks search empty state
