+++ modules/features/player/src/main/java/au/com/shiftyjelly/pocketcasts/player/view/bookmark/BookmarksFragment.kt
@@ -145,9 +145,6 @@ class BookmarksFragment : BaseFragment() {
                             addFragment(fragment)
                         }
                     },
-                    onClearSearchTapped = {
-                        bookmarksViewModel.clearSearchTapped()
-                    },
                     onSearchBarClearButtonTapped = {
                         bookmarksViewModel.searchBarClearButtonTapped()
                     },
