+++ modules/features/player/src/main/java/au/com/shiftyjelly/pocketcasts/player/view/bookmark/BookmarksPage.kt
@@ -42,7 +42,7 @@ import au.com.shiftyjelly.pocketcasts.models.entity.Bookmark
 import au.com.shiftyjelly.pocketcasts.models.entity.PodcastEpisode
 import au.com.shiftyjelly.pocketcasts.models.type.SyncStatus
 import au.com.shiftyjelly.pocketcasts.player.view.bookmark.components.HeaderRow
-import au.com.shiftyjelly.pocketcasts.player.view.bookmark.components.NoBookmarksInSearchView
+import au.com.shiftyjelly.pocketcasts.player.view.bookmark.components.NoMatchingBookmarks
 import au.com.shiftyjelly.pocketcasts.player.viewmodel.BookmarksViewModel
 import au.com.shiftyjelly.pocketcasts.player.viewmodel.BookmarksViewModel.BookmarkMessage
 import au.com.shiftyjelly.pocketcasts.player.viewmodel.BookmarksViewModel.UiState
@@ -70,7 +70,6 @@ fun BookmarksPage(
     onUpgradeClicked: () -> Unit,
     showOptionsDialog: (Int) -> Unit,
     openFragment: (Fragment) -> Unit,
-    onClearSearchTapped: () -> Unit,
     onSearchBarClearButtonTapped: () -> Unit,
     onHeadphoneControlsButtonTapped: () -> Unit,
     bottomInset: Dp,
@@ -93,7 +92,6 @@ fun BookmarksPage(
         onUpgradeClicked = onUpgradeClicked,
         openFragment = openFragment,
         bottomInset = bottomInset,
-        onClearSearchTapped = onClearSearchTapped,
         onSearchBarClearButtonTapped = onSearchBarClearButtonTapped,
         onHeadphoneControlsButtonTapped = onHeadphoneControlsButtonTapped,
         isDarkTheme = isDarkTheme,
@@ -143,7 +141,6 @@ private fun Content(
     onSearchTextChanged: (String) -> Unit,
     onUpgradeClicked: () -> Unit,
     openFragment: (Fragment) -> Unit,
-    onClearSearchTapped: () -> Unit,
     onSearchBarClearButtonTapped: () -> Unit,
     onHeadphoneControlsButtonTapped: () -> Unit,
     bottomInset: Dp,
@@ -166,7 +163,6 @@ private fun Content(
                 onPlayClick = onPlayClick,
                 onSearchTextChanged = onSearchTextChanged,
                 bottomInset = bottomInset,
-                onClearSearchTapped = onClearSearchTapped,
                 onSearchBarClearButtonTapped = onSearchBarClearButtonTapped,
                 isDarkTheme = isDarkTheme,
             )
@@ -208,16 +204,15 @@ private fun BookmarksView(
     onOptionsMenuClicked: () -> Unit,
     onPlayClick: (Bookmark) -> Unit,
     onSearchTextChanged: (String) -> Unit,
-    onClearSearchTapped: () -> Unit,
     onSearchBarClearButtonTapped: () -> Unit,
     bottomInset: Dp,
     isDarkTheme: Boolean,
 ) {
     val focusRequester = remember { FocusRequester() }
     LazyColumn(
+        horizontalAlignment = Alignment.CenterHorizontally,
         contentPadding = PaddingValues(bottom = bottomInset),
-        modifier = Modifier
-            .fillMaxSize(),
+        modifier = Modifier.fillMaxSize(),
     ) {
         if (state.searchEnabled) {
             item {
@@ -240,11 +235,8 @@ private fun BookmarksView(
             state.bookmarks.isEmpty()
         ) {
             item {
-                NoBookmarksInSearchView(
-                    onActionClick = {
-                        onClearSearchTapped()
-                        onSearchTextChanged("")
-                    },
+                NoMatchingBookmarks(
+                    modifier = Modifier.padding(top = 24.dp),
                 )
             }
         } else {
@@ -336,7 +328,6 @@ private fun BookmarksPreview(
             onSearchTextChanged = {},
             onUpgradeClicked = {},
             openFragment = {},
-            onClearSearchTapped = {},
             onSearchBarClearButtonTapped = {},
             onHeadphoneControlsButtonTapped = {},
             bottomInset = 0.dp,
